refactor: deduplicate executeSave into shared save-actions utility
Extracted the duplicated save-project logic from 3 components
(accelerator-handler, file menu, activity bar) and 2 modals
(save-changes-file, save-changes-project) into a single
executeSaveProject() function in save-actions.ts.

Also exported sanitizePou, collectDebugVariables, and EditorLike
from save-project.ts for reuse by the upcoming single-file save.

Added docs/unified-frontend-serialization.md with the full plan
for moving parsing and serialization to the frontend.
